Rename ReadStateStoreCommitterLogs
