+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/statestore/ReadStateStoreCommitterLogsTest.java
@@ -17,15 +17,15 @@ package sleeper.systemtest.drivers.statestore;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.systemtest.drivers.statestore.StateStoreCommitterLogEntry.LambdaFinished;
-import sleeper.systemtest.drivers.statestore.StateStoreCommitterLogEntry.LambdaStarted;
+import sleeper.systemtest.drivers.statestore.ReadStateStoreCommitterLogs.LambdaFinished;
+import sleeper.systemtest.drivers.statestore.ReadStateStoreCommitterLogs.LambdaStarted;
 import sleeper.systemtest.dsl.statestore.StateStoreCommitSummary;
 
 import java.time.Instant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class StateStoreCommitterLogEntryTest {
+public class ReadStateStoreCommitterLogsTest {
 
     @Test
     void shouldReadLambdaStarted() {
@@ -33,7 +33,7 @@ public class StateStoreCommitterLogEntryTest {
         String message = "[main] committer.lambda.StateStoreCommitterLambda INFO - Lambda started at 2024-08-13T12:12:00Z";
 
         // When / Then
-        assertThat(StateStoreCommitterLogEntry.readEvent("test-stream", message)).isEqualTo(
+        assertThat(ReadStateStoreCommitterLogs.readEvent("test-stream", message)).isEqualTo(
                 new LambdaStarted(Instant.parse("2024-08-13T12:12:00Z")));
     }
 
@@ -43,7 +43,7 @@ public class StateStoreCommitterLogEntryTest {
         String message = "[main] committer.lambda.StateStoreCommitterLambda INFO - Lambda finished at 2024-08-13T12:13:00Z (ran for 1 minute)";
 
         // When / Then
-        assertThat(StateStoreCommitterLogEntry.readEvent("test-stream", message)).isEqualTo(
+        assertThat(ReadStateStoreCommitterLogs.readEvent("test-stream", message)).isEqualTo(
                 new LambdaFinished(Instant.parse("2024-08-13T12:13:00Z")));
     }
 
@@ -53,7 +53,7 @@ public class StateStoreCommitterLogEntryTest {
         String message = "[main] sleeper.commit.StateStoreCommitter INFO - Applied request to table ID test-table with type TestRequest at time 2024-08-13T12:12:30Z";
 
         // When / Then
-        assertThat(StateStoreCommitterLogEntry.readEvent("test-stream", message)).isEqualTo(
+        assertThat(ReadStateStoreCommitterLogs.readEvent("test-stream", message)).isEqualTo(
                 new StateStoreCommitSummary("test-stream", "test-table", "TestRequest", Instant.parse("2024-08-13T12:12:30Z")));
     }
 
@@ -63,6 +63,6 @@ public class StateStoreCommitterLogEntryTest {
         String message = "some other log";
 
         // When / Then
-        assertThat(StateStoreCommitterLogEntry.readEvent("test-stream", message)).isNull();
+        assertThat(ReadStateStoreCommitterLogs.readEvent("test-stream", message)).isNull();
     }
 }
